[IMP] point_of_sale, product, website_sale: migrate base unit logic
This commit moves the base unit fields used in `website_sale`'s
`product.template` and `product.product` up to the `product` module
in order to allow other inheritors (such as `point_of_sale`) to use
them. The setting to display it was also added to `point_of_sale`
so that `website_sale` doesn't have to be installed to enable it.

Task ID: 4277792
