+++ addons/point_of_sale/views/res_config_settings_views.xml
@@ -286,6 +286,10 @@
                             <setting id="pos-loyalty" title="Boost your sales with multiple kinds of programs: Coupons, Promotions, Gift Card, Loyalty. Specific conditions can be set (products, customers, minimum purchase amount, period). Rewards can be discounts (% or amount) or free products." string="Promotions, Coupons, Gift Card &amp; Loyalty Program" help="Manage promotion that will grant customers discounts or gifts" documentation="/applications/sales/point_of_sale/pricing/loyalty.html">
                                 <field name="module_loyalty"/>
                             </setting>
+                            <setting id="pos_uom_price_option_setting" string="Product Reference Price"
+                                     help="Add a reference price per UoM on products (i.e $/kg), in addition to the sale price">
+                                <field name="group_show_uom_price"/>
+                            </setting>
                         </block>
 
                         <block title="Bills &amp; Receipts" id="pos_bills_and_receipts_section">
