+++ addons/product/security/ir.model.access.csv
@@ -37,3 +37,6 @@ access_update_product_attribute_value_manager,update.product.attribute.value,mod
 access_product_uom_manager,product.uom manager,model_product_uom,group_product_manager,1,1,1,1
 access_product_uom_user,product.uom user,model_product_uom,base.group_user,1,0,0,0
 access_uom_uom_product_manager,uom.uom.product.manager,model_uom_uom,group_product_manager,1,1,1,1
+access_product_base_unit_public_public,product.base.unit public,model_product_base_unit,base.group_public,1,0,0,0
+access_product_base_unit_public_portal,product.base.unit public,model_product_base_unit,base.group_portal,1,0,0,0
+access_product_base_unit_public_employee,product.base.unit public,model_product_base_unit,base.group_user,1,0,0,0
