+++ addons/product/views/product_base_unit_views.xml
@@ -1,10 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo>
-
     <record id="base_unit_action" model="ir.actions.act_window">
         <field name="name">Base Units</field>
-        <field name="res_model">website.base.unit</field>
+        <field name="res_model">product.base.unit</field>
         <field name="view_mode">list,form</field>
     </record>
-
 </odoo>
