+++ addons/sale/security/ir.model.access.csv
@@ -21,6 +21,7 @@ access_product_pricelist_sale_user,product.pricelist.sale.user,product.model_pro
 access_product_pricelist_sale_manager,product.pricelist salemanager,product.model_product_pricelist,sales_team.group_sale_manager,1,1,1,1
 access_product_pricelist_item_sale_manager,product.pricelist.item salemanager,product.model_product_pricelist_item,sales_team.group_sale_manager,1,1,1,1
 access_product_document_sale_manager,access_product_document_sale_manager,product.model_product_document,sales_team.group_sale_manager,1,1,1,1
+access_product_base_unit_sale_manager,product.base.unit sale manager,product.model_product_base_unit,sales_team.group_sale_manager,1,1,1,1
 
 access_uom_uom_user,uom.uom.user,uom.model_uom_uom,sales_team.group_sale_salesman,1,0,0,0
