+++ addons/website_sale/models/__init__.py
@@ -24,7 +24,6 @@ from . import res_partner
 from . import sale_order
 from . import sale_order_line
 from . import website
-from . import website_base_unit
 from . import website_checkout_step
 from . import website_menu
 from . import website_sale_extra_field
