+++ addons/website_sale/models/product_product.py
@@ -23,28 +23,6 @@ class ProductProduct(models.Model):
         inverse_name='product_variant_id',
     )
 
-    base_unit_count = fields.Float(
-        string="Base Unit Count",
-        help="Display base unit price on your eCommerce pages. Set to 0 to hide it for this"
-             " product.",
-        required=True,
-        default=1,
-    )
-    base_unit_id = fields.Many2one(
-        string="Custom Unit of Measure",
-        help="Define a custom unit to display in the price per unit of measure field.",
-        comodel_name='website.base.unit',
-    )
-    base_unit_price = fields.Monetary(
-        string="Price Per Unit",
-        compute='_compute_base_unit_price',
-    )
-    base_unit_name = fields.Char(
-        help="Displays the custom unit for the products if defined or the selected unit of measure"
-            " otherwise.",
-        compute='_compute_base_unit_name',
-    )
-
     website_url = fields.Char(
         string="Website URL",
         help="The full URL to access the document through the website.",
@@ -52,23 +30,6 @@ class ProductProduct(models.Model):
     )
 
     #=== COMPUTE METHODS ===#
-
-    def _get_base_unit_price(self, price):
-        self.ensure_one()
-        return self.base_unit_count and price / self.base_unit_count
-
-    @api.depends('lst_price', 'base_unit_count')
-    def _compute_base_unit_price(self):
-        for product in self:
-            if not product.id:
-                product.base_unit_price = 0
-            else:
-                product.base_unit_price = product._get_base_unit_price(product.lst_price)
-
-    @api.depends('uom_name', 'base_unit_id')
-    def _compute_base_unit_name(self):
-        for product in self:
-            product.base_unit_name = product.base_unit_id.name or product.uom_name
 
     @api.depends_context('lang')
     @api.depends('product_tmpl_id.website_url', 'product_template_attribute_value_ids')
@@ -80,22 +41,7 @@ class ProductProduct(models.Model):
                 url = f'{url}#attribute_values={",".join(pav_ids)}'
             product.website_url = url
 
-    #=== CONSTRAINT METHODS ===#
-
-    @api.constrains('base_unit_count')
-    def _check_base_unit_count(self):
-        if any(product.base_unit_count < 0 for product in self):
-            raise ValidationError(_(
-                "The value of Base Unit Count must be greater than 0."
-                " Use 0 to hide the price per unit on this product."
-            ))
-
     #=== BUSINESS METHODS ===#
-
-    def _prepare_variant_values(self, combination):
-        variant_dict = super()._prepare_variant_values(combination)
-        variant_dict['base_unit_count'] = self.base_unit_count
-        return variant_dict
 
     def website_publish_button(self):
         self.ensure_one()
